requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
boto3==1.26.137
botocore==1.29.137
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect2==0.7.0
dj-database-url==2.0.0
dj-static==0.0.6
Django==4.1.7
django-discover-runner==1.0
django-extensions==3.2.1
django-mathfilters==1.0.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-role-permissions==3.1.1
django-smart-selects==1.6.0
django-storages==1.13.2
django-widget-tweaks==1.4.12
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
jsonfield==3.1.0
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pydot==1.4.2
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pyparsing==3.0.9
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
svglib==1.5.1
swapper==1.3.0
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
